ADD: Restructure with runtimes/ and utils/
